Refactored the project to clear up any naming scheme issues.
-Controllers call on managers
-managers retrieve the object/inject the object from memory and ES
