+++ app/src/main/java/ca/ualberta/cs/cmput301f18t19/hada/hada/ui/ProblemListActivity.java
@@ -34,7 +34,7 @@ public class ProblemListActivity extends AppCompatActivity {
             }
         });
 
-        //Goes to EditUserSettingsActivityroge
+        //Goes to EditUserSettingsActivity
         ImageButton settingsButton = findViewById(R.id.problemListSettingsButton);
         settingsButton.setOnClickListener(new View.OnClickListener() {
             @Override
