+++ app/src/test/java/ca/ualberta/cs/cmput301f18t19/hada/hada/model/PatientControllerTest.java
@@ -3,21 +3,21 @@ package ca.ualberta.cs.cmput301f18t19.hada.hada.model;
 import org.junit.Test;
 
 import ca.ualberta.cs.cmput301f18t19.hada.hada.model.Patient;
-import ca.ualberta.cs.cmput301f18t19.hada.hada.model.PatientListController;
+import ca.ualberta.cs.cmput301f18t19.hada.hada.model.PatientController;
 
 import static org.junit.Assert.*;
 
-public class PatientListControllerTest {
+public class PatientControllerTest {
 
     @Test
     public void testIsEmpty() {
-        PatientListController patientList = new PatientListController();
+        PatientController patientList = new PatientController();
         assertTrue(patientList.isEmpty());
     }
 
     @Test
     public void testAddPatient() {
-        PatientListController patientList = new PatientListController();
+        PatientController patientList = new PatientController();
         Patient patient = new Patient();
         assertTrue(patientList.isEmpty());
         patientList.addPatient(patient);
@@ -26,7 +26,7 @@ public class PatientListControllerTest {
 
     @Test
     public void testGetPatient() {
-        PatientListController patientList = new PatientListController();
+        PatientController patientList = new PatientController();
         Patient patient = new Patient();
         patientList.addPatient(patient);
         assertEquals(patient, patientList.getPatient(0));
@@ -34,7 +34,7 @@ public class PatientListControllerTest {
 
     @Test
     public void testGetSize() {
-        PatientListController patientList = new PatientListController();
+        PatientController patientList = new PatientController();
         Patient patient = new Patient();
         assertEquals(0, patientList.getSize());
         patientList.addPatient(patient);
@@ -43,7 +43,7 @@ public class PatientListControllerTest {
 
     @Test
     public void testPatientInList() {
-        PatientListController patientList = new PatientListController();
+        PatientController patientList = new PatientController();
         Patient patient1 = new Patient();
         Patient patient2 = new Patient();
         assertFalse(patientList.inList(patient1));
@@ -55,7 +55,7 @@ public class PatientListControllerTest {
 
     @Test
     public void testGetPosition() {
-        PatientListController patientList = new PatientListController();
+        PatientController patientList = new PatientController();
         Patient patient1 = new Patient();
         Patient patient2 = new Patient();
         Patient patient3 = new Patient();
@@ -68,7 +68,7 @@ public class PatientListControllerTest {
 
     @Test
     public void testDeletePatient() {
-        PatientListController patientList = new PatientListController();
+        PatientController patientList = new PatientController();
         Patient patient1 = new Patient();
         Patient patient2 = new Patient();
         Patient patient3 = new Patient();
@@ -82,7 +82,7 @@ public class PatientListControllerTest {
 
     @Test
     public void testInsertPatient() {
-        PatientListController patientList = new PatientListController();
+        PatientController patientList = new PatientController();
         Patient patient1 = new Patient();
         Patient patient2 = new Patient();
         Patient patient3 = new Patient();
